refactor: split large mapper/converter files to comply with 100-line limit
- Split Zwift target.mapper.ts (175 lines) into:
  - zwift-to-krd.mapper.ts (Zwift → KRD conversions)
  - krd-to-zwift.mapper.ts (KRD → Zwift conversions)
  - target.mapper.ts (re-exports for backward compatibility)

- Split Zwift target.converter.ts (135 lines) into:
  - power.converter.ts (power-related conversions)
  - pace-cadence.converter.ts (pace and cadence conversions)
  - target.converter.ts (re-exports for backward compatibility)

- Split FIT target-power.converter.ts (124 lines) into:
  - power-helpers.ts (helper functions)
  - target-power.converter.ts (main conversion logic)

- Split Zwift intervals-t.mapper.ts (116 lines) into:
  - intervals-t-helpers.ts (createOnStep, createOffStep)
  - intervals-t.mapper.ts (main mapper function)

- Split TCX tcx-to-krd.converter.ts (111 lines) into:
  - standard-duration.converter.ts (standard duration types)
  - extended-duration.converter.ts (extended duration types)
  - tcx-to-krd.converter.ts (main converter with re-exports)

All changes maintain backward compatibility through re-exports.
All tests pass (735/736 passing).
Coverage remains at 92.64%.
