+++ packages/core/src/adapters/tcx/duration/tcx-to-krd.converter.ts
@@ -1,96 +1,17 @@
+import { durationTypeSchema } from "../../../domain/schemas/duration";
+import { tcxDurationTypeSchema } from "../schemas/tcx-duration";
+import { convertExtendedDuration } from "./extended-duration.converter";
 import {
-  durationTypeSchema,
-  type Duration,
-} from "../../../domain/schemas/duration";
-import {
-  tcxDurationTypeSchema,
-  type TcxDurationType,
-} from "../schemas/tcx-duration";
+  convertStandardDuration,
+  type TcxDurationConversionResult,
+  type TcxDurationData,
+} from "./standard-duration.converter";
 
-export type TcxDurationData = {
-  durationType?: string;
-  seconds?: number;
-  meters?: number;
-  bpm?: number;
-  calories?: number;
-};
-
-export type TcxDurationExtensions = {
-  heartRateAbove?: number;
-  heartRateBelow?: number;
-  caloriesBurned?: number;
-};
-
-export type TcxDurationConversionResult = {
-  duration: Duration;
-  extensions?: TcxDurationExtensions;
-};
-
-const convertStandardDuration = (
-  durationType: TcxDurationType,
-  data: TcxDurationData
-): TcxDurationConversionResult | null => {
-  if (
-    durationType === tcxDurationTypeSchema.enum.Time &&
-    data.seconds !== undefined
-  ) {
-    return {
-      duration: { type: durationTypeSchema.enum.time, seconds: data.seconds },
-    };
-  }
-
-  if (
-    durationType === tcxDurationTypeSchema.enum.Distance &&
-    data.meters !== undefined
-  ) {
-    return {
-      duration: { type: durationTypeSchema.enum.distance, meters: data.meters },
-    };
-  }
-
-  if (durationType === tcxDurationTypeSchema.enum.LapButton) {
-    return { duration: { type: durationTypeSchema.enum.open } };
-  }
-
-  return null;
-};
-
-const convertExtendedDuration = (
-  durationType: TcxDurationType,
-  data: TcxDurationData
-): TcxDurationConversionResult | null => {
-  if (
-    durationType === tcxDurationTypeSchema.enum.HeartRateAbove &&
-    data.bpm !== undefined
-  ) {
-    return {
-      duration: { type: durationTypeSchema.enum.open },
-      extensions: { heartRateAbove: data.bpm },
-    };
-  }
-
-  if (
-    durationType === tcxDurationTypeSchema.enum.HeartRateBelow &&
-    data.bpm !== undefined
-  ) {
-    return {
-      duration: { type: durationTypeSchema.enum.open },
-      extensions: { heartRateBelow: data.bpm },
-    };
-  }
-
-  if (
-    durationType === tcxDurationTypeSchema.enum.CaloriesBurned &&
-    data.calories !== undefined
-  ) {
-    return {
-      duration: { type: durationTypeSchema.enum.open },
-      extensions: { caloriesBurned: data.calories },
-    };
-  }
-
-  return null;
-};
+export type {
+  TcxDurationConversionResult,
+  TcxDurationData,
+  TcxDurationExtensions,
+} from "./standard-duration.converter";
 
 export const convertTcxDuration = (
   data: TcxDurationData
